# Locallaplacian speedup 2