--- a/src/common/locallaplacian.c
+++ b/src/common/locallaplacian.c
@@ -29,6 +29,8 @@
 #include <xmmintrin.h>
 #endif
 
+//#define DEBUG_DUMP
+
 // downsample width/height to given level
 static inline int dl(int size, const int level)
 {
@@ -36,6 +38,22 @@ static inline int dl(int size, const int level)
     size = (size-1)/2+1;
   return size;
 }
+
+#ifdef DEBUG_DUMP
+static void dump_PFM(const char *filename, const float* out, const uint32_t w, const uint32_t h)
+{
+  FILE *f = g_fopen(filename, "wb");
+  fprintf(f, "PF\n%d %d\n-1.0\n", w, h);
+  for(int j=0;j<h;j++)
+    for(int i=0;i<w;i++)
+      for(int c=0;c<3;c++)
+        fwrite(out + w*j+i, 1, sizeof(float), f);
+  fclose(f);
+}
+#define debug_dump_PFM dump_PFM
+#else
+#define debug_dump_PFM(f,b,w,h)
+#endif
 
 // needs a boundary of 1 or 2px around i,j or else it will crash.
 // (translates to a 1px boundary around the corresponding pixel in the coarse buffer)
@@ -348,20 +366,15 @@ static inline float *ll_pad_input(
     }
     pad_by_replication(out, *wd2, *ht2, max_supp);
   }
-#if 0
+#ifdef DEBUG_DUMP
   if(b && b->mode == 2)
   {
-    FILE *f = fopen("/tmp/padded.pfm", "wb");
-    fprintf(f, "PF\n%d %d\n-1.0\n", *wd2, *ht2);
-    for(int j=0;j<*ht2;j++)
-      for(int i=0;i<*wd2;i++)
-        for(int c=0;c<3;c++)
-          fwrite(out + *wd2*j+i, 1, sizeof(float), f);
-    fclose(f);
+    dump_PFM("/tmp/padded.pfm",out,*wd2,*ht2);
   }
 #endif
   return out;
 }
+
 
 static inline float ll_laplacian(
     const float *const coarse,   // coarse res gaussian
@@ -640,28 +653,8 @@ void local_laplacian_internal(
     const int pw = dl(w,last_level), ph = dl(h,last_level);
     const int pw0 = dl(b->pwd, pl0), ph0 = dl(b->pht, pl0);
     const int pw1 = dl(b->pwd, pl1), ph1 = dl(b->pht, pl1);
-#if 0
-    {
-    FILE *f = fopen("/tmp/coarse.pfm", "wb");
-    fprintf(f, "PF\n%d %d\n-1.0\n", pw0, ph0);
-    for(int j=0;j<ph0;j++)
-      for(int i=0;i<pw0;i++)
-        for(int c=0;c<3;c++)
-          fwrite(b->output[pl0] + pw0*j+i, 1, sizeof(float), f);
-    fclose(f);
-    }
-#endif
-#if 0
-    {
-    FILE *f = fopen("/tmp/oldcoarse.pfm", "wb");
-    fprintf(f, "PF\n%d %d\n-1.0\n", pw, ph);
-    for(int j=0;j<ph;j++)
-      for(int i=0;i<pw;i++)
-        for(int c=0;c<3;c++)
-          fwrite(output[last_level] + pw*j+i, 1, sizeof(float), f);
-    fclose(f);
-    }
-#endif
+    debug_dump_PFM("/tmp/coarse.pfm", b->output[pl0], pw0, ph0);
+    debug_dump_PFM("/tmp/oldcoarse.pfm", output[last_level], pw, ph);
 #ifdef _OPENMP
 #pragma omp parallel for schedule(static) collapse(2) default(shared)
 #endif
@@ -699,17 +692,7 @@ void local_laplacian_internal(
 #endif
       output[last_level][j*pw+i] = weight * c1 + (1.0f-weight) * c0;
     }
-#if 0
-    {
-    FILE *f = fopen("/tmp/newcoarse.pfm", "wb");
-    fprintf(f, "PF\n%d %d\n-1.0\n", pw, ph);
-    for(int j=0;j<ph;j++)
-      for(int i=0;i<pw;i++)
-        for(int c=0;c<3;c++)
-          fwrite(output[last_level] + pw*j+i, 1, sizeof(float), f);
-    fclose(f);
-    }
-#endif
+    debug_dump_PFM("/tmp/newcoarse.pfm", output[last_level], pw, ph);
   }
 
   // assemble output pyramid coarse to fine