By rearranging the order of computation, I was able to make the gauss_reduce_sse2 function in src/common/locallaplacian.c parallelizable.  Each thread can now be given 1/t of the entire 2d array instead of 1/t of a single row.  Being able to use vector instructions yields a 20+% speedup compared to the scalar version used in the current Git master.

|		| Git master		| New code		|
| ---		| ---			| ---			|
| No OpenMP	|3.695 secs (3.829 CPU)	|2.447 secs (2.674 CPU)	|
|2 threads	|1.940 secs (3.421 CPU)	|1.477 secs (2.476 CPU)	|
|8 threads	|0.627 secs (4.034 CPU)	|0.451 secs (2.691 CPU)	|
|32 threads	|0.302 secs (6.694 CPU)	|0.262 secs (4.072 CPU)	|
|64 threads	|0.373 secs (7.803 CPU)	|0.280 secs (5.905 CPU)	|

In the process, I found that the scalar and SSE versions were using different kernels.  The third commit in this PR adapts the scalar version to use the SSE version's kernel, since I figure the vast majority of users will have been running that code.

Scalar and SSE2 versions now produce the same output modulo floating-point roundoff causing some of the pixel values to differ by 1 (of a range of 65536) in 16-bit output.

Fixes #4808.