@johnny-bit 2020-04-25
In #4154 @burrima added unit testing capacities. Cases like this are perfect use-case for those. Can you try and also create unit tests for locallaplacian code?