@ralfbrown 2020-04-25
I'll take a look  (from scanning that PR, it seems I'll have to learn both Cmake and cmocka to do so).  I'm also not sure how one would make unit tests for something whose results depend on the entire input image....

For cases like this, it may be better to have a system test.  Several representative images that a test script runs through darktable-cli and compares against the "gold" result.  For the comparison that I referred to above, what I did was export to full-resolution 16-bit TIFFs using both the pre- and post-change versions of dt (did it manually though), and then used ImageMagick to compare the resulting images.

~~~
  convert before.tif after.tif -metric pae -compare -format "%[distortion]\n" info:
~~~
should give a result of either 0 (all pixels the same) or 1.52959e-05 (1/65536, differences due to roundoff).  If there were differences, you can count the number of differing pixels with
~~~
  convert before.tif after.tif -metric ae -compare -format "%[distortion]\n" info:
~~~

Speaking of testing, can someone tell me why Travis CI fails the PR?  All I get when I try to see the details is a blank page.