@TurboGit 2020-04-25
Oh, I have already a test for local contrast in fact, let me test this PR against it...