@TurboGit 2020-04-25
Result is not good:

```
$ ./run.sh 0002-local-contrast
Test 0002-local-contrast
      Image mire1.cr2
      CPU & GPU version differs
      Max  dE         11
      Mean dE         0
      Count above max 88474
  FAILS: image visually changed
         see diff.jpg for visual difference

Total test 1
Errors     1
```
Here is the resulting diff image:

![diff](https://user-images.githubusercontent.com/467069/80286860-36736900-872e-11ea-822a-f565bb4827fc.jpg)