@ralfbrown 2020-04-25
Which versions of the code did you compare for that difference image?  Keep in mind that my first-round speedup changed from gauss_reduce_sse2 to gauss_reduce, which as I later discovered is using a different kernel.  This PR changes things to use the same kernel that was in use before my first PR.

Since gauss_expand has factors of 4 and 6 in the computation, I think the 1 4 6 4 1 kernel is in fact the correct one, not the 1 5 8 5 1 that gauss_reduce() had been using.