@TurboGit 2020-04-25
> Which versions of the code did you compare for that difference image?

The version from some (~20) minutes ago.