@ralfbrown 2020-04-25
What version for the reference image?  The correct comparison is against the code prior to my first PR for local contrast (#4789), because that PR inadvertently changed the results due to the inconsistency between the then-existing scalar and SSE code.