@TurboGit 2020-04-25
BTW, there is certainly an issue of non initialized value as I get different results on each run of the testsuite.