@TurboGit 2020-04-25
> What version for the reference image?

The reference image has been created just before merging this issue with up-to-date code from master.