@ralfbrown 2020-04-26
But up-to-date code from master includes my previous PR, which breaks things.... (by using the 1 5 8 5 1 kernel in the scalar code, instead of the 1 4 6 4 1 kernel used by both SSE and CL code)  This PR fixes that by making every code path use the 1 4 6 4 1 kernel, but that in itself changes the output if you only compare with and without this single PR.

I'll look for uninitialized values.