@ralfbrown 2020-04-26
Found it -- off-by-one error.  Multiple consecutive exports give identical results now.