@TurboGit 2020-04-26
Ok, I had indeed created the reference from your previous commit. I just recreated a proper reference image, and now my test gives on current master:

```
Test 0002-local-contrast
      Image mire1.cr2
      CPU & GPU version differs
      Max  dE         11
      Mean dE         0
      Count above max 88469
  FAILS: image visually changed
         see diff.jpg for visual difference
```

With this PR it is a bit better but not perfect:

```
Test 0002-local-contrast
      Image mire1.cr2
      CPU & GPU version differs
      Max  dE         3
      Mean dE         0
      Count above max 69
  FAILS: image visually changed
         see diff.jpg for visual difference


Total test 1
Errors     1
```

Here is the diff and the good news is that this time the diff is stable so no more random pixels.

![diff](https://user-images.githubusercontent.com/467069/80300547-09f73580-879e-11ea-90f9-6e0cb012f7bf.jpg)

Can you please see if the remaining errors can be fixed?

Thanks.