@ralfbrown 2020-04-26
Unable to reproduce using TIFF output.

Your use of JPEG output is creating compression artifacts that amplify differences to the point that the minimum possible difference results in an artifact exceeding the dE threshold.  I recognized that test image, so I went and downloaded the .CR2 for Canon 5D Mark III from imaging-resource (hard to find a CR2, since they don't link superceded cameras from their index).

Using
~~~
darktable-cli --hq true --apply-custom-presets false E5D3hSLI000800NR0.CR2{,.xmp} output.tif
~~~
ImageMagick with
~~~
convert expected.tif output.tif -metric pae -compare info:
~~~
reports a maximal pixel difference of 1.5259e-05, which is 1/65536, the minimal representable difference in a 16 bit per channel image.  Change all of the .tif above to .jpg, and the PAE jumps to 0.0235294.