@TurboGit 2020-04-26
I fully agree that jpeg is not a good choice. I've used that because I found some sample Python code to read image to be passed to OpenCV to compute Delta-E. I'm sure we can use PNG or TIF.