@TurboGit 2020-04-26
All good now ! I have migrated the testsuite to use PNG and the diff compared to the reference version is Delta-E 1. Will merge now.